# public/swagger.yaml
openapi: 3.0.4
info:
  title: API for WMS - Documented by Swagger
  description: |-
    This is an API documentation of all endpoints available for the use of WMS

  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
  /api/customer-order:
    get:
      tags:
        - Outbound functions
      summary: Get all available customer orders
      description: Get all information (CustomerOrderId, OrderDate, Status,...) abount all available customer orders
      operationId: getAllCustomerOrders
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerOrder"

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Outbound functions
      summary: Add a new customer order to WMS.
      description: Insert a new customer order record into the system.
      operationId: addCustomerOrder
      requestBody:
        description: Create a new customer order
        content:
          application/json:
            $ref: "#/components/schemas/InsertCustomerOrder"
        required: true
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrder"

        "500":
          description: Error inserting data
          schema:
            $ref: "#/components/schemas/Error"
    put:
      tags:
        - Outbound functions
      summary: Update the status of the customer order.
      description: All other field of customer order should be static (out of scope of a WMS). Only status is allowed to be editted.
      operationId: updateCustomerOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerOrder"
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrder"

        "500":
          description: Error updating data
          schema:
            $ref: "#/components/schemas/Error"
...
  /api/inbound-shipment:
    get:
      tags:
        - Inbound functions
      summary: Get all available customer orders
      description: Get all information about all available inbound shipments
      operationId: getAllInboundShipment
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboundShipment"

        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Inbound functions
      summary: Add an inbound shipment to WMS.
      description: Create a new customer order. There is only 2 fields in the inbound shipment. Both can be auto-generated by the database. Hence, no need for body request.
      operationId: addInboundShipment
      responses:
        "201":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundShipment"

        "500":
          description: Error inserting data
          schema:
            $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Inbound functions
      summary: Delete an inbound shipment by shipId
      operationId: deleteInboundShipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipId:
                  type: integer
                  example: 101
              required:
                - shipId
      responses:
        "200":
          description: Inbound shipment deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundShipment"
        "400":
          description: Invalid or missing shipId
        "404":
          description: No shipment found with given ID
        "500":
          description: Internal server error during deletion
  /api/inbound-shipment/{shipId}:
    get:
      tags:
        - Inbound functions
      summary: Get the details of the inbound shipment by its shipId.
      description: Get the details of products within the inbound shipment idetified by the shipId.
      operationId: getInboundShipmentDetails
      parameters:
        - name: shipId
          in: path
          description: The key ID used to find target shipment.
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        "200":
          description: successful fetching data of inbound shipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundShipmentDetails"

        "404":
          description: Invalid orderId value
          example: No order found with ID 10003
        "500":
          description: Error fetching data
    post:
      tags:
        - Inbound functions
      summary: Manually insert a new Inbound Shipment Detail record
      description: Typically handled by a trigger, but this endpoint allows manual insertion for testing or manual correction.
      operationId: createInboundShipmentDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipId:
                  type: integer
                  example: 101
                productId:
                  type: integer
                  example: 2005
                receivedQuantity:
                  type: integer
                  minimum: 0
                  example: 0
              required:
                - shipId
                - productId
                - receivedQuantity
      responses:
        "201":
          description: Inbound shipment detail successfully created
          content:
            application/json:
              schema:
                type: object
                description: Result of the database insert operation
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ship ID, Product ID and Received Quantity are required
        "500":
          description: Server error while inserting data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error inserting data
    put:
      tags:
        - Inbound functions
      summary: Update received quantity for a specific inbound shipment detail
      description: Updates the `receivedQuantity` for a specific combination of `shipId` and `productId`.
      operationId: updateInboundShipmentDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipId:
                  type: integer
                  example: 101
                productId:
                  type: integer
                  example: 2005
                receivedQuantity:
                  type: integer
                  minimum: 0
                  example: 25
              required:
                - shipId
                - productId
                - receivedQuantity
      responses:
        "200":
          description: Received quantity successfully updated
          content:
            application/json:
              schema:
                type: object
                description: Result of the update operation (e.g., affected row count or raw result)
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Ship ID, Product ID and Received Quantity are required
        "500":
          description: Server error while updating data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error updating data
...
  /api/order-transaction:
    get:
      tags:
        - Stock functions
      summary: Get the information of all order transactions
      description: This is all the transactions (operations) performed on the the order.
      operationId: getOrderTransactions
      responses:
        "200":
        description: Successfully fetching the information of all order transactions
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/OrderTransaction"
    post:
      tags:
        - Stock functions
      summary: Create a new order transaction
      description: Creates a new transaction record for a customer order. Note that transactions of type `Receive` or ones missing `refId` are handled by triggers and should not be inserted via this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customerOrderId
                - transactionType
              properties:
                customerOrderId:
                  type: integer
                  example: 123
                  description: The ID of the associated customer order
                refId:
                  type: integer
                  nullable: true
                  example: 456
                  description: A reference ID for the transaction (thiss ref ID can be shipID ~ outbound or pickListID)
                transactionType:
                  type: string
                  enum:
                    - Pick and Pack
                    - Ship
                  example: "Pick and Pack"
                  description: The type of transaction. Only allowing inserting Pick and Pack or Ship transactions. "Receive" transaction is automated.
                transactionTime:
                  type: string
                  format: date-time
                  nullable: true
                  example: "2025-04-24T12:34:56Z"
                  description: Optional custom timestamp for the transaction
      responses:
        "201":
          description: Order transaction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Order transaction created successfully.
                  data:
                    type: object
                    description: Insert result or record metadata
        "400":
          description: Bad request due to missing or invalid fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: customerOrderId and transactionType are required.
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error while creating order transaction.
...
components:
  schemas:
    CustomerOrder:
      type: object
      properties:
        customerOrderID:
          description: "The primary key of CustomerOrder"
          oneOf:
            - type: string
            - type: integer
          example: 1001
        orderDate:
          description: "The date the order was placed"
          type: string
          format: date-time
          example: "2025-04-17 11:05:38"
        status:
          description: "The current status of the order"
          type: string
          enum:
            - Pending
            - Pick and Pack
            - Shipped
          example: "Pick and"
        address:
          description: "The address where the order should be placed"
          type: string
          example: "456 Oak Ave, Somewhere, SM 67890"
    UpdateCustomerOrder:
      type: object
      required:
        - orderId
        - newStatus
      properties:
        orderId:
          description: "The key to search for the target customer order for updating."
          oneOf:
            - type: string
            - type: integer

          example: 1001

        newStatus:
          description: "The new status of the updating customer order."
          type: string
          enum:
            - Pending
            - Pick and Pack
            - Shipped
          example: "Pick and Pack"
    InsertCustomerOrder:
      type: object
      required:
        - address
      properties:
        orderDate:
          description: "The date the order was placed, if not given, take current timestamp as default"
          type: string
          format: date-time
          example: "2025-04-17 11:05:38"

        address:
          description: "The address where the order should be placed"
          type: string
          example: "456 Oak Ave, Somewhere, SM 67890"
    InboundShipment:
      type: object
      properties:
        shipID:
          oneOf:
            - type: string
            - type: integer
          description: The key of inbound shipment.
        shipmentTime:
          type: string
          format: date-time
          description: The date the shipment is inputted. If not given, take current timestamp as default.
    InboundShipmentDetails:
      type: object
      properties:
        shipId:
          oneOf:
            - type: string
            - type: integer
          description: The key of inbound shipment
        inboundShipmentDetails:
          type: object
          properties:
            productId:
              oneOf:
                - type: string
                - type: integer
              description: Foreign key referencing Product.productId.
            receivedQuantity:
              type: integer
              minimum: 0
              description: Quantity received. Must be >= 0.
              example: 50
            orderedQuantity:
              type: integer
              minimum: 0
              description: Quantity ordered in PO. Must be >= 0.
              example: 55
        purchaseOrderIds:
          type: array
          items:
            type: integer
          example: [1001, 1002, 1003]
...
    OrderTransaction:
      type: object
      required:
        - transactionId
        - customerOrderId
        - transactionType
      properties:
        transactionId:
          type: integer
          format: int32
          description: Auto-incremented primary key
          example: 1
        customerOrderId:
          type: integer
          format: int32
          description: Foreign key referencing CustomerOrder.customerOrderId
          example: 101
        refId:
          type: integer
          format: int32
          nullable: true
          description: Optional reference ID
          example: 5001
        transactionTime:
          type: string
          format: date-time
          description: Timestamp of the transaction
          example: "2025-04-24T15:30:00Z"
        transactionType:
          type: string
          enum:
            - Receive
            - Pick and Pack
            - Ship
          description: Type of transaction
          example: "Receive"
...
    Error:
      type: object
      properties:
        status:
          type: string
          example: 500
        message:
          type: string
          example: "Unexpected Error"
      required:
        - status
        - message